moved repetitive html into constants file
